User mask
This can be used as alternative way to mask a user's avatar with that typical shout-it shape. I had problems with the SVG clipping mask because the positioning with css, negative margins, etc :( This should be a more solid solution.

See how it is used in HeaderProfile.css. If we need it in other places, we could even make a small dumb component instead. We could make webpack to load the image as data-uri if we want to improve its rendering (now while it's loading the image, you may still see the not-masked image).