Fix store rehydratation
This commit bring some sanity to the flux rehydrating part, required to fix #218 and #208. I've been careful to pass to store initializers the right data, but it was very confusing and I hope I didn't broke anything.

While cleaning it up, I could see web.js it is rehydrating stores using a `data` object which is used to initialize many stores. What's going into this object is not clear and  its content should become explicit when we improve our server-side renderings.

There was also a `router` variable that wasn't never used to stores (this is correct, stores shouldn't be aware of the router), but for some reasons it was included in the variable initialization.

/cc @p0o