Better styling for textareas
@p0o this is a veeery first step to reviewing our stylesheet.
The problem here is that there are many selectors without class. Here I'm removing the one for the textarea. I hope there are no big consequences.

As convention, I'm thinking we could use `.html<TagName>` for marking classes that should be attached to html tags. Not sure yet if the best choice, for now I chose this because it easy to refactor :)

This kind of classes goes into html.scss.